# More efficient encoding of case classes