--- a/zio-json/shared/src/main/scala-3/zio/json/macros.scala
+++ b/zio-json/shared/src/main/scala-3/zio/json/macros.scala
@@ -601,7 +601,11 @@ sealed class JsonEncoderDerivation(config: JsonCodecConfiguration) extends Deriv
           while (idx < fields.length) {
             val field = fields(idx)
             val p     = field.p.deref(a)
-            field.encodeOrSkip(p) { () =>
+            val encoder = field.encoder
+            if ((field.withExplicitNulls && field.withExplicitEmptyCollections) ||
+              (field.withExplicitNulls && !encoder.isEmpty(p)) ||
+              (field.withExplicitEmptyCollections && !encoder.isNothing(p)) ||
+              (!encoder.isEmpty(p) && !encoder.isNothing(p))) {
               // if we have at least one field already, we need a comma
               if (prevFields) {
                 out.write(',')
@@ -610,7 +614,7 @@ sealed class JsonEncoderDerivation(config: JsonCodecConfiguration) extends Deriv
               JsonEncoder.string.unsafeEncode(field.name, indent_, out)
               if (indent.isEmpty) out.write(':')
               else out.write(" : ")
-              field.encoder.unsafeEncode(p, indent_, out)
+              encoder.unsafeEncode(p, indent_, out)
               prevFields = true // at least one field so far
             }
             idx += 1