--- a/zio-json/shared/src/main/scala-3/zio/json/macros.scala
+++ b/zio-json/shared/src/main/scala-3/zio/json/macros.scala
@@ -602,10 +602,14 @@ sealed class JsonEncoderDerivation(config: JsonCodecConfiguration) extends Deriv
             val field = fields(idx)
             val p     = field.p.deref(a)
             val encoder = field.encoder
-            if ((field.withExplicitNulls && field.withExplicitEmptyCollections) ||
-              (field.withExplicitNulls && !encoder.isEmpty(p)) ||
-              (field.withExplicitEmptyCollections && !encoder.isNothing(p)) ||
-              (!encoder.isEmpty(p) && !encoder.isNothing(p))) {
+            if ({
+              (field.flags: @switch) match {
+                case 0 => !encoder.isEmpty(p) && !encoder.isNothing(p)
+                case 1 => !encoder.isNothing(p)
+                case 2 => !encoder.isEmpty(p)
+                case _ => true
+              }
+            }) {
               // if we have at least one field already, we need a comma
               if (prevFields) {
                 out.write(',')