--- a/zio-json/shared/src/main/scala-3/zio/json/macros.scala
+++ b/zio-json/shared/src/main/scala-3/zio/json/macros.scala
@@ -577,7 +577,7 @@ sealed class JsonEncoderDerivation(config: JsonCodecConfiguration) extends Deriv
           val withExplicitEmptyCollections = p.annotations.collectFirst { case a: jsonExplicitEmptyCollections =>
             a.encoding
           }.getOrElse(explicitEmptyCollections)
-          new FieldEncoder(
+          FieldEncoder(
             p,
             name,
             p.typeclass.asInstanceOf[JsonEncoder[Any]],