--- a/zio-json/shared/src/test/scala/zio/json/internal/FieldEncoderHelperSpec.scala
+++ b/zio-json/shared/src/test/scala/zio/json/internal/FieldEncoderHelperSpec.scala
@@ -10,7 +10,7 @@ object FieldEncoderSpec extends ZIOSpecDefault {
     suite("encodeOrDefault")(
       suite("OptionEncoder")(
         test("should use the default encoding None when withExplicitNulls is false") {
-          val helper = new FieldEncoder(
+          val helper = FieldEncoder(
             1,
             "test",
             JsonEncoder.option(JsonEncoder.int),
@@ -23,7 +23,7 @@ object FieldEncoderSpec extends ZIOSpecDefault {
           )
         },
         test("should encode None when withExplicitNulls is true") {
-          val helper = new FieldEncoder(
+          val helper = FieldEncoder(
             1,
             "test",
             JsonEncoder.option(JsonEncoder.int),
@@ -38,7 +38,7 @@ object FieldEncoderSpec extends ZIOSpecDefault {
       ),
       suite("CollectionEncoder")(
         test("should encode empty collections when withExplicitEmptyCollections is true") {
-          val helper = new FieldEncoder(
+          val helper = FieldEncoder(
             1,
             "test",
             implicitly[JsonEncoder[List[Int]]],
@@ -51,7 +51,7 @@ object FieldEncoderSpec extends ZIOSpecDefault {
           )
         },
         test("should not encode empty collections when withExplicitEmptyCollections is false") {
-          val helper = new FieldEncoder(
+          val helper = FieldEncoder(
             1,
             "test",
             implicitly[JsonEncoder[List[Int]]],
@@ -67,7 +67,7 @@ object FieldEncoderSpec extends ZIOSpecDefault {
       suite("for a case class")(
         test("should encode case classes with empty collections when withExplicitEmptyCollections is true") {
           case class Test(list: List[Int], option: Option[Int])
-          val helper = new FieldEncoder(
+          val helper = FieldEncoder(
             1,
             "test",
             DeriveJsonEncoder.gen[Test],
@@ -84,7 +84,7 @@ object FieldEncoderSpec extends ZIOSpecDefault {
         },
         test("should not encode case classes with empty collections when withExplicitEmptyCollections is false") {
           case class Test(list: List[Int], option: Option[Int])
-          val helper = new FieldEncoder(
+          val helper = FieldEncoder(
             1,
             "test",
             DeriveJsonEncoder.gen[Test],
@@ -100,7 +100,7 @@ object FieldEncoderSpec extends ZIOSpecDefault {
           "should also not encode case classes with empty options when withExplicitEmptyCollections is false, even when withExplicitNulls is true"
         ) {
           case class Test(list: List[Int], option: Option[Int])
-          val helper = new FieldEncoder(
+          val helper = FieldEncoder(
             1,
             "test",
             DeriveJsonEncoder.gen[Test],