Tested with Scala 3 and JDK-21 on  Intel® Core™ i7-11800H:

#### Before
```txt
[info] Benchmark                           Mode  Cnt       Score      Error  Units
[info] GeoJSONBenchmarks.encodeZio        thrpt    5   23442.369 ±  882.412  ops/s
[info] GoogleMapsAPIBenchmarks.encodeZio  thrpt    5   31923.719 ±  519.186  ops/s
[info] SyntheticBenchmarks.encodeZio      thrpt    5  359053.440 ± 4165.845  ops/s
[info] TwitterAPIBenchmarks.encodeZio     thrpt    5   43622.904 ± 1043.312  ops/s
```

#### After
```txt
[info] Benchmark                           Mode  Cnt       Score      Error  Units
[info] GeoJSONBenchmarks.encodeZio        thrpt    5   23634.989 ±  534.950  ops/s
[info] GoogleMapsAPIBenchmarks.encodeZio  thrpt    5   39406.055 ± 1003.726  ops/s
[info] SyntheticBenchmarks.encodeZio      thrpt    5  522796.537 ± 2469.604  ops/s
[info] TwitterAPIBenchmarks.encodeZio     thrpt    5   50215.699 ± 1517.542  ops/s
```